含めたくない要素を指定できるようにした

+++ mkf_element.hpp
@@ -24,7 +24,7 @@ namespace mkf{
 
 
 enum class
-ElementKind: uint16_t
+ElementKind: uint8_t
 {
   null,
   string,
